refactor(format): split serializer/parser into their feature folders
Move serializeProjectTo8f4e into project-export/serializeTo8f4e.ts
and parse8f4eToProject into project-import/parse8f4e.ts.
Remove pureHelpers/projectFormat8f4e.ts. Update all import paths.
